requirements.txt
Pillow==10.2.0 	
streamlit==1.22.0
pandas==2.0.1
scikit_learn==1.2.2
numpy==1.24.3
